update-V1

+++ se.gu.uoy.dsl/src/patterns/AtLeast.java
@@ -25,12 +25,12 @@ public interface AtLeast extends ElementaryPatterns {
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Value</em>' attribute.
-	 * @see #setValue(int)
+	 * @see #setValue(double)
 	 * @see patterns.PatternsPackage#getAtLeast_Value()
 	 * @model required="true"
 	 * @generated
 	 */
-	int getValue();
+	double getValue();
 
 	/**
 	 * Sets the value of the '{@link patterns.AtLeast#getValue <em>Value</em>}' attribute.
@@ -40,6 +40,6 @@ public interface AtLeast extends ElementaryPatterns {
 	 * @see #getValue()
 	 * @generated
 	 */
-	void setValue(int value);
+	void setValue(double value);
 
 } // AtLeast
