+++ se.gu.uoy.dsl/src/patterns/StrictlyWithin.java
@@ -26,12 +26,12 @@ public interface StrictlyWithin extends ElementaryPatterns {
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Value1</em>' attribute.
-	 * @see #setValue1(int)
+	 * @see #setValue1(double)
 	 * @see patterns.PatternsPackage#getStrictlyWithin_Value1()
 	 * @model required="true"
 	 * @generated
 	 */
-	int getValue1();
+	double getValue1();
 
 	/**
 	 * Sets the value of the '{@link patterns.StrictlyWithin#getValue1 <em>Value1</em>}' attribute.
@@ -41,19 +41,19 @@ public interface StrictlyWithin extends ElementaryPatterns {
 	 * @see #getValue1()
 	 * @generated
 	 */
-	void setValue1(int value);
+	void setValue1(double value);
 
 	/**
 	 * Returns the value of the '<em><b>Value2</b></em>' attribute.
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Value2</em>' attribute.
-	 * @see #setValue2(int)
+	 * @see #setValue2(double)
 	 * @see patterns.PatternsPackage#getStrictlyWithin_Value2()
 	 * @model required="true"
 	 * @generated
 	 */
-	int getValue2();
+	double getValue2();
 
 	/**
 	 * Sets the value of the '{@link patterns.StrictlyWithin#getValue2 <em>Value2</em>}' attribute.
@@ -63,6 +63,6 @@ public interface StrictlyWithin extends ElementaryPatterns {
 	 * @see #getValue2()
 	 * @generated
 	 */
-	void setValue2(int value);
+	void setValue2(double value);
 
 } // StrictlyWithin
